Move CanonicalHeaders into novoda package
