+++ java-library/src/main/java/uk/co/lucasweb/aws/v4/signer/CanonicalRequest.java
@@ -12,9 +12,10 @@
  */
 package uk.co.lucasweb.aws.v4.signer;
 
+import com.novoda.aws.v4.signer.CanonicalHeaders;
+
 import java.net.URI;
 import java.net.URISyntaxException;
-
 import java.util.ArrayList;
 import java.util.Comparator;
 import java.util.List;
